requirements.txt
streamlit==1.17.0
streamlit-authenticator==0.1.5
numpy==1.23.5
opencv-python-headless
pandas==1.5.3
Pillow==9.4.0
opencv-python
datetime
